requirements.txt
flask>=2.0.0
flask-login>=0.5.0
requests>=2.26.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
email-validator>=2.0.0
openai>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5 